% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_plot.R
\name{create_parameter_plot}
\alias{create_parameter_plot}
\title{Create a plot comparing a single parameter from multiple sources, such as multiple scenarios.}
\usage{
create_parameter_plot(scenariosFrame, yAxisTitle)
}
\arguments{
\item{scenariosFrame}{A "wide" data frame with a 'Time' column and a series of each source/scenario's values.}

\item{yAxisTitle}{The label to provide the y-axis (describe the parameter visualized across sources)}
}
\value{
ggplot plot with each source/scenario as its own series and a legend mapping series to color.
}
\description{
Create a plot comparing a single parameter from multiple sources, such as multiple scenarios.
}
