% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_scenarios.R
\name{visualize_scenarios}
\alias{visualize_scenarios}
\title{Draws a multi-panel graphic of the key parameters of a SWMM model across multiple design scenarios.}
\usage{
visualize_scenarios(
  outputFiles,
  outputScenarioNames,
  storageNodes,
  controlledLinks,
  passiveLinks,
  title,
  flowUnits = "CFS"
)
}
\arguments{
\item{outputFiles}{List of paths to SWMM '.out' output files with data to visualize. Each scenario being compared should have 1 output file.}

\item{outputScenarioNames}{Labels for the time-series of each scenario. This list must be in the same order as outputFiles.}

\item{storageNodes}{List of node (Storage Units, Junctions, Outfalls, and Dividers) names to include in visualization. Storage nodes are placed at the top of the graphic.}

\item{controlledLinks}{List of links (Conduits, Pumps, Orifices, Weirs, and Outlets) that were actively controlled in at least one of the scenarios.
The graphic presents both flow rate and capacity for all Controlled links.}

\item{passiveLinks}{List of links (Conduits, Pumps, Orifices, Weirs, and Outlets) that were not controlled in any scenario but should be visualized nonetheless
(i.e. to show impact of upstream controls). Passive links are displayed at the bottom of the graphic.}

\item{title}{The graphic title to use.}

\item{flowUnits}{The flow units used by the SWMM model underlying these scenario results, either CFS or CMS.}
}
\description{
Draws a multi-panel graphic of the key parameters of a SWMM model across multiple design scenarios.
}
\examples{
visualize_scenarios(
  outputFiles = c("/dev/swmmscenario/inst/theta.out", "/dev/swmmscenario/inst/theta_optimizedsmartretention.out"),
  outputScenarioNames = c("Passive Baseline", "Smart Retention"),
  storageNodes = c("P1"),
  controlledLinks = c("1"),
  passiveLinks = c("8"),
  title = "Theta 50yr Design Storm",
  flowUnits = "CMS")
}
